yazi config

+++ scripts/stow.sh
@@ -4,10 +4,11 @@ sudo pacman -S --noconfirm --needed stow
 
 echo "Running stow..."
 
-#stow doesnt override files if they already exist so we have to move existing ones
-mv ~/.bashrc ~/.bashrc.backup
-mv ~/.config/nvim ~/.config/nvim.backup
-mv ~/.config/ghostty/config ~/.config/ghostty/config.backup
-mv ~/.config/starship.toml ~/.config/starship.toml.backup
+# stow doesnt override files if they already exist so we have to move existing ones
+# it only moves files if destination doesnt already exist (only move at first run)
+[ ! -e ~/.bashrc.backup ] && mv ~/.bashrc ~/.bashrc.backup
+[ ! -e ~/.config/nvim.backup ] && mv ~/.config/nvim ~/.config/nvim.backup
+[ ! -e ~/.config/ghostty/config.backup ] && mv ~/.config/ghostty/config ~/.config/ghostty/config.backup
+[ ! -e ~/.config/starship.toml.backup ] && mv ~/.config/starship.toml ~/.config/starship.toml.backup
 
 stow -t ~ configs
